bridge-node: wire SomaRelayer + expose BridgeState via proto + fix toml u128 deser (testnet-v0.1.24-rc1)
Three fixes needed for the testnet bridge to operate end-to-end:

1. proto: add `bridge_state` to `SystemState` + nested `BridgeState`,
   `BridgeCommittee`, `BridgeMember`, `BridgeRegistration` messages, plus
   read/write conversions in `rpc_proto_conversions.rs`. Replaces the
   hardcoded `BridgeState::new(BridgeCommittee::empty())` so callers of
   `Client::get_latest_system_state` (notably `bridge-node`) see the live
   committee + registrations instead of an empty stub.

2. bridge-node: `WatchdogConfigBlock.in_flight_tolerance_micro` u128->u64.
   The `toml` crate errors with "u128 is not supported" on deserialize,
   which crashed bridge nodes whenever the optional `[watchdog]` section
   was present. u64 covers the addressable range (1.8e13 USDC); the
   comparison in `check_once` is widened to u128 at the comparison site.

3. bridge-node: surface a `[soma_relayer]` config block + thread it
   through `BridgeNode::with_relayer` in `main.rs`. Previously the node
   had no way to assemble + submit quorum certs back to Soma — it ran
   sig-cache-only. End-to-end bridge transfers require this wiring.

Also bumps the workspace to 0.1.24 to invalidate the cargo-chef cache for
the bridge-node Cloud Build image, and includes the local
`DeployBridge.s.sol` parse fix from earlier today (the JSON
`somaCommitteeDigest` 0x-prefixed string was auto-decoded by foundry as
`bytes32`, breaking `abi.decode(..., (string))`).

+++ bridge-node/src/watchdog.rs
@@ -401,7 +401,9 @@ pub struct ConservationInvariantObservable {
     pub failure_threshold: u32,
     /// Tolerated USDC delta in microdollars. Covers transient
     /// in-flight transfer volume. Zero requires exact equality.
-    pub in_flight_tolerance_micro: u128,
+    /// Widened to `u128` only at the comparison site (see
+    /// `check_once`) so the field stays TOML-deserializable.
+    pub in_flight_tolerance_micro: u64,
     /// Signing queue handle — auto-pause posts an EmergencyPause
     /// action here when the violation threshold trips.
     pub signing_tx: mpsc::Sender<BridgeActionExecutionWrapper>,
@@ -424,7 +426,7 @@ impl ConservationInvariantObservable {
         eth_bridge_contract_address: String,
         interval: Duration,
         failure_threshold: u32,
-        in_flight_tolerance_micro: u128,
+        in_flight_tolerance_micro: u64,
         signing_tx: mpsc::Sender<BridgeActionExecutionWrapper>,
         expected_pause_nonce: ExpectedPauseNonceReader,
     ) -> Self {
@@ -453,7 +455,7 @@ impl ConservationInvariantObservable {
             .await?;
         let soma_supply = (self.soma_supply)().await? as u128;
         info!(eth_locked, soma_supply, "ConservationInvariant: paired reading");
-        let limit = eth_locked.saturating_add(self.in_flight_tolerance_micro);
+        let limit = eth_locked.saturating_add(self.in_flight_tolerance_micro as u128);
         Ok(soma_supply <= limit)
     }
